+++ mne/viz/epochs.py
@@ -22,7 +22,7 @@ from ..time_frequency import psd_multitaper
 from .utils import (tight_layout, figure_nobar, _toggle_proj, _toggle_options,
                     _layout_figure, _setup_vmin_vmax, _channels_changed,
                     _plot_raw_onscroll, _onclick_help, plt_show,
-                    _compute_scalings)
+                    _compute_scalings, DraggableColorbar)
 from ..defaults import _handle_default
 
 
@@ -65,8 +65,15 @@ def plot_epochs_image(epochs, picks=None, sigma=0., vmin=None,
         The scalings of the channel types to be applied for plotting.
         If None, defaults to `scalings=dict(eeg=1e6, grad=1e13, mag=1e15,
         eog=1e6)`.
-    cmap : matplotlib colormap
-        Colormap.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive'
+        Colormap. If tuple, the first value indicates the colormap to use and
+        the second value is a boolean defining interactivity. In interactive
+        mode the colors are adjustable by clicking and dragging the colorbar
+        with left and right mouse button. Left mouse button moves the scale up
+        and down and right mouse button adjusts the range. Hitting space bar
+        resets the scale. Up and down arrows can be used to change the
+        colormap. If 'interactive', translates to ('RdBu_r', True). Defaults to
+        'RdBu_r'.
     fig : matplotlib figure | None
         Figure instance to draw the image to. Figure must contain two axes for
         drawing the single trials and evoked responses. If None a new figure is
@@ -176,15 +183,19 @@ def plot_epochs_image(epochs, picks=None, sigma=0., vmin=None,
             ax2 = plt.subplot2grid((3, 10), (2, 0), colspan=9, rowspan=1)
             if colorbar:
                 ax3 = plt.subplot2grid((3, 10), (0, 9), colspan=1, rowspan=3)
-        if scale_vmin:
-            vmin *= scalings[ch_type]
-        if scale_vmax:
-            vmax *= scalings[ch_type]
+
+        this_vmin = vmin * scalings[ch_type] if scale_vmin else vmin
+        this_vmax = vmax * scalings[ch_type] if scale_vmax else vmax
+
+        if cmap == 'interactive':
+            cmap = ('RdBu_r', True)
+        elif not isinstance(cmap, tuple):
+            cmap = (cmap, True)
         im = ax1.imshow(this_data,
                         extent=[1e3 * epochs.times[0], 1e3 * epochs.times[-1],
                                 0, len(data)],
                         aspect='auto', origin='lower', interpolation='nearest',
-                        vmin=vmin, vmax=vmax, cmap=cmap)
+                        vmin=this_vmin, vmax=this_vmax, cmap=cmap[0])
         if this_overlay_times is not None:
             plt.plot(1e3 * this_overlay_times, 0.5 + np.arange(len(this_data)),
                      'k', linewidth=2)
@@ -206,10 +217,12 @@ def plot_epochs_image(epochs, picks=None, sigma=0., vmin=None,
         ax2.set_ylim([evoked_vmin, evoked_vmax])
         ax2.axvline(0, color='m', linewidth=3, linestyle='--')
         if colorbar:
-            plt.colorbar(im, cax=ax3)
+            cbar = plt.colorbar(im, cax=ax3)
+            if cmap[1]:
+                ax1.CB = DraggableColorbar(cbar, im)
             tight_layout(fig=this_fig)
-
     plt_show(show)
+
     return figs