+++ mne/viz/tests/test_evoked.py
@@ -116,12 +116,12 @@ def test_plot_evoked():
 
         evoked.plot_image(proj=True)
         # plot with bad channels excluded
-        evoked.plot_image(exclude='bads')
+        evoked.plot_image(exclude='bads', cmap='interactive')
         evoked.plot_image(exclude=evoked.info['bads'])  # does the same thing
         plt.close('all')
 
         evoked.plot_topo()  # should auto-find layout
-        _butterfly_onselect(0, 200, ['mag'], evoked)  # test averaged topomap
+        _butterfly_onselect(0, 200, ['mag', 'grad'], evoked)
         plt.close('all')
 
         cov = read_cov(cov_fname)