+++ mne/viz/topo.py
@@ -21,7 +21,8 @@ from ..utils import _clean_names, warn
 from ..channels.layout import _merge_grad_data, _pair_grad_sensors, find_layout
 from ..defaults import _handle_default
 from .utils import (_check_delayed_ssp, COLORS, _draw_proj_checkbox,
-                    add_background_image, plt_show, _setup_vmin_vmax)
+                    add_background_image, plt_show, _setup_vmin_vmax,
+                    DraggableColorbar)
 
 
 def iter_topography(info, layout=None, on_pick=None, fig=None,
@@ -251,13 +252,15 @@ def _check_vlim(vlim):
 
 def _imshow_tfr(ax, ch_idx, tmin, tmax, vmin, vmax, onselect, ylim=None,
                 tfr=None, freq=None, vline=None, x_label=None, y_label=None,
-                colorbar=False, picker=True, cmap='RdBu_r', title=None,
+                colorbar=False, picker=True, cmap=('RdBu_r', True), title=None,
                 hline=None):
     """ Aux function to show time-freq map on topo """
     import matplotlib.pyplot as plt
     from matplotlib.widgets import RectangleSelector
 
     extent = (tmin, tmax, freq[0], freq[-1])
+    cmap, interactive_cmap = cmap
+
     img = ax.imshow(tfr[ch_idx], extent=extent, aspect="auto", origin="lower",
                     vmin=vmin, vmax=vmax, picker=picker, cmap=cmap)
     if isinstance(ax, plt.Axes):
@@ -271,7 +274,12 @@ def _imshow_tfr(ax, ch_idx, tmin, tmax, vmin, vmax, onselect, ylim=None,
         if y_label is not None:
             plt.ylabel(y_label)
     if colorbar:
-        plt.colorbar(mappable=img)
+        if isinstance(colorbar, DraggableColorbar):
+            cbar = colorbar.cbar  # this happens with multiaxes case
+        else:
+            cbar = plt.colorbar(mappable=img)
+        if interactive_cmap:
+            ax.CB = DraggableColorbar(cbar, img)
     if title:
         plt.title(title)
     if not isinstance(ax, plt.Axes):