+++ mne/viz/topomap.py
@@ -23,7 +23,7 @@ from ..io.pick import (pick_types, _picks_by_type, channel_type, pick_info,
 from ..utils import _clean_names, _time_mask, verbose, logger, warn
 from .utils import (tight_layout, _setup_vmin_vmax, _prepare_trellis,
                     _check_delayed_ssp, _draw_proj_checkbox, figure_nobar,
-                    plt_show, _process_times)
+                    plt_show, _process_times, DraggableColorbar)
 from ..time_frequency import psd_multitaper
 from ..defaults import _handle_default
 from ..channels.layout import _find_topomap_coords
@@ -135,9 +135,16 @@ def plot_projs_topomap(projs, layout=None, cmap=None, sensors=True,
         Layout instance specifying sensor positions (does not need to be
         specified for Neuromag data). Or a list of Layout if projections
         are from different sensor types.
-    cmap : matplotlib colormap | None
-        Colormap to use. If None, 'Reds' is used for all positive data,
-        otherwise defaults to 'RdBu_r'.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode (only works if ``colorbar=True``) the colors are
+        adjustable by clicking and dragging the colorbar with left and right
+        mouse button. Left mouse button moves the scale up and down and right
+        mouse button adjusts the range. Hitting space bar resets the range. Up
+        and down arrows can be used to change the colormap. If None (default),
+        'Reds' is used for all positive data, otherwise defaults to 'RdBu_r'.
+        If 'interactive', translates to (None, True).
     sensors : bool | str
         Add markers for sensor locations to the plot. Accepts matplotlib plot
         format string (e.g., 'r+' for red plusses). If True, a circle will be
@@ -182,7 +189,7 @@ def plot_projs_topomap(projs, layout=None, cmap=None, sensors=True,
     .. versionadded:: 0.9.0
     """
     import matplotlib.pyplot as plt
-
+    from mpl_toolkits.axes_grid1 import make_axes_locatable
     if layout is None:
         from ..channels import read_layout
         layout = read_layout('Vectorview-all')
@@ -194,6 +201,10 @@ def plot_projs_topomap(projs, layout=None, cmap=None, sensors=True,
     nrows = math.floor(math.sqrt(n_projs))
     ncols = math.ceil(n_projs / nrows)
 
+    if cmap == 'interactive':
+        cmap = (None, True)
+    elif not isinstance(cmap, tuple):
+        cmap = (cmap, True)
     if axes is None:
         plt.figure()
         axes = list()
@@ -232,12 +243,16 @@ def plot_projs_topomap(projs, layout=None, cmap=None, sensors=True,
             break
 
         if len(idx):
-            plot_topomap(data, pos[:, :2], vmax=None, cmap=cmap,
-                         sensors=sensors, res=res, axes=axes[proj_idx],
-                         outlines=outlines, contours=contours,
-                         image_interp=image_interp, show=False)
+            im = plot_topomap(data, pos[:, :2], vmax=None, cmap=cmap[0],
+                              sensors=sensors, res=res, axes=axes[proj_idx],
+                              outlines=outlines, contours=contours,
+                              image_interp=image_interp, show=False)[0]
             if colorbar:
-                plt.colorbar()
+                divider = make_axes_locatable(axes[proj_idx])
+                cax = divider.append_axes("right", size="5%", pad=0.05)
+                cbar = plt.colorbar(im, cax=cax, cmap=cmap)
+                if cmap[1]:
+                    axes[proj_idx].CB = DraggableColorbar(cbar, im)
         else:
             raise RuntimeError('Cannot find a proper layout for projection %s'
                                % proj['desc'])
@@ -785,8 +800,20 @@ def plot_ica_components(ica, picks=None, ch_type=None, res=64,
         The value specifying the upper bound of the color range.
         If None, the maximum absolute value is used. If callable, the output
         equals vmax(data). Defaults to None.
-    cmap : matplotlib colormap
-        Colormap.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode the colors are adjustable by clicking and dragging the
+        colorbar with left and right mouse button. Left mouse button moves the
+        scale up and down and right mouse button adjusts the range. Hitting
+        space bar resets the range. Up and down arrows can be used to change
+        the colormap. If None, 'Reds' is used for all positive data,
+        otherwise defaults to 'RdBu_r'. If 'interactive', translates to
+        (None, True). Defaults to 'RdBu_r'.
+
+        .. warning::  Interactive mode works smoothly only for a small amount
+            of topomaps.
+
     sensors : bool | str
         Add markers for sensor locations to the plot. Accepts matplotlib
         plot format string (e.g., 'r+' for red plusses). If True, a circle
@@ -835,9 +862,9 @@ def plot_ica_components(ica, picks=None, ch_type=None, res=64,
         figs = []
         for k in range(0, n_components, p):
             picks = range(k, min(k + p, n_components))
-            fig = plot_ica_components(ica, picks=picks,
-                                      ch_type=ch_type, res=res, layout=layout,
-                                      vmax=vmax, cmap=cmap, sensors=sensors,
+            fig = plot_ica_components(ica, picks=picks, ch_type=ch_type,
+                                      res=res, layout=layout, vmax=vmax,
+                                      cmap=cmap, sensors=sensors,
                                       colorbar=colorbar, title=title,
                                       show=show, outlines=outlines,
                                       contours=contours,
@@ -848,6 +875,15 @@ def plot_ica_components(ica, picks=None, ch_type=None, res=64,
         picks = [picks]
     ch_type = _get_ch_type(ica, ch_type)
 
+    if cmap == 'interactive':
+        cmap = ('RdBu_r', True)
+    elif not isinstance(cmap, tuple):
+        if len(picks) > 2:
+            warn('Disabling interactive colorbar for multiple axes. Turn '
+                 'interactivity on explicitly by passing cmap as a tuple.')
+            cmap = (cmap, False)
+        else:
+            cmap = (cmap, True)
     data = np.dot(ica.mixing_matrix_[:, picks].T,
                   ica.pca_components_[:ica.n_components_])
 
@@ -879,7 +915,7 @@ def plot_ica_components(ica, picks=None, ch_type=None, res=64,
         data_ = _merge_grad_data(data_) if merge_grads else data_
         vmin_, vmax_ = _setup_vmin_vmax(data_, vmin, vmax)
         im = plot_topomap(data_.flatten(), pos, vmin=vmin_, vmax=vmax_,
-                          res=res, axes=ax, cmap=cmap, outlines=outlines,
+                          res=res, axes=ax, cmap=cmap[0], outlines=outlines,
                           image_mask=image_mask, contours=contours,
                           image_interp=image_interp, show=False)[0]
         if colorbar:
@@ -889,6 +925,8 @@ def plot_ica_components(ica, picks=None, ch_type=None, res=64,
             cbar.ax.tick_params(labelsize=12)
             cbar.set_ticks((vmin_, vmax_))
             cbar.ax.set_title('AU', fontsize=10)
+            if cmap[1]:
+                ax.CB = DraggableColorbar(cbar, im)
         _hide_frame(ax)
     tight_layout(fig=fig)
     fig.subplots_adjust(top=0.95)
@@ -954,10 +992,16 @@ def plot_tfr_topomap(tfr, tmin=None, tmax=None, fmin=None, fmax=None,
         The value specifying the upper bound of the color range. If None, the
         maximum value is used. If callable, the output equals vmax(data).
         Defaults to None.
-    cmap : matplotlib colormap | None
-        Colormap. If None and the plotted data is all positive, defaults to
-        'Reds'. If None and data contains also negative values, defaults to
-        'RdBu_r'. Defaults to None.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode the colors are adjustable by clicking and dragging the
+        colorbar with left and right mouse button. Left mouse button moves the
+        scale up and down and right mouse button adjusts the range. Hitting
+        space bar resets the range. Up and down arrows can be used to change
+        the colormap. If None (default), 'Reds' is used for all positive data,
+        otherwise defaults to 'RdBu_r'. If 'interactive', translates to
+        (None, True).
     sensors : bool | str
         Add markers for sensor locations to the plot. Accepts matplotlib
         plot format string (e.g., 'r+' for red plusses). If True, a circle will
@@ -1044,8 +1088,10 @@ def plot_tfr_topomap(tfr, tmin=None, tmax=None, fmin=None, fmax=None,
 
     norm = False if np.min(data) < 0 else True
     vmin, vmax = _setup_vmin_vmax(data, vmin, vmax, norm)
-    if cmap is None:
-        cmap = 'Reds' if norm else 'RdBu_r'
+    if cmap is None or cmap == 'interactive':
+        cmap = ('Reds', True) if norm else ('RdBu_r', True)
+    elif not isinstance(cmap, tuple):
+        cmap = (cmap, True)
 
     if axes is None:
         fig = plt.figure()
@@ -1061,21 +1107,23 @@ def plot_tfr_topomap(tfr, tmin=None, tmax=None, fmin=None, fmax=None,
     fig_wrapper = list()
     selection_callback = partial(_onselect, tfr=tfr, pos=pos, ch_type=ch_type,
                                  itmin=itmin, itmax=itmax, ifmin=ifmin,
-                                 ifmax=ifmax, cmap=cmap, fig=fig_wrapper,
+                                 ifmax=ifmax, cmap=cmap[0], fig=fig_wrapper,
                                  layout=layout)
 
     im, _ = plot_topomap(data[:, 0], pos, vmin=vmin, vmax=vmax,
-                         axes=ax, cmap=cmap, image_interp='bilinear',
+                         axes=ax, cmap=cmap[0], image_interp='bilinear',
                          contours=False, names=names, show_names=show_names,
                          show=False, onselect=selection_callback)
 
     if colorbar:
         divider = make_axes_locatable(ax)
         cax = divider.append_axes("right", size="5%", pad=0.05)
-        cbar = plt.colorbar(im, cax=cax, format=cbar_fmt, cmap=cmap)
+        cbar = plt.colorbar(im, cax=cax, format=cbar_fmt, cmap=cmap[0])
         cbar.set_ticks((vmin, vmax))
         cbar.ax.tick_params(labelsize=12)
         cbar.ax.set_title('AU')
+        if cmap[1]:
+            ax.CB = DraggableColorbar(cbar, im)
 
     plt_show(show)
     return fig
@@ -1119,9 +1167,20 @@ def plot_evoked_topomap(evoked, times="auto", ch_type=None, layout=None,
         The value specifying the upper bound of the color range.
         If None, the maximum absolute value is used. If callable, the output
         equals vmax(data). Defaults to None.
-    cmap : matplotlib colormap | None
-        Colormap to use. If None, 'Reds' is used for all positive data,
-        otherwise defaults to 'RdBu_r'.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode the colors are adjustable by clicking and dragging the
+        colorbar with left and right mouse button. Left mouse button moves the
+        scale up and down and right mouse button adjusts the range. Hitting
+        space bar resets the range. Up and down arrows can be used to change
+        the colormap. If None (default), 'Reds' is used for all positive data,
+        otherwise defaults to 'RdBu_r'. If 'interactive', translates to
+        (None, True).
+
+        .. warning::  Interactive mode works smoothly only for a small amount
+            of topomaps.
+
     sensors : bool | str
         Add markers for sensor locations to the plot. Accepts matplotlib plot
         format string (e.g., 'r+' for red plusses). If True, a circle will be
@@ -1317,10 +1376,19 @@ def plot_evoked_topomap(evoked, times="auto", ch_type=None, layout=None,
              for i in range(len(times))]
     vmin = np.min(vlims)
     vmax = np.max(vlims)
+    if cmap == 'interactive':
+        cmap = (None, True)
+    elif not isinstance(cmap, tuple):
+        if len(times) > 2:
+            warn('Disabling interactive colorbar for multiple axes. Turn '
+                 'interactivity on explicitly by passing cmap as a tuple.')
+            cmap = (cmap, False)
+        else:
+            cmap = (cmap, True)
     for idx, time in enumerate(times):
         tp, cn = plot_topomap(data[:, idx], pos, vmin=vmin, vmax=vmax,
                               sensors=sensors, res=res, names=names,
-                              show_names=show_names, cmap=cmap,
+                              show_names=show_names, cmap=cmap[0],
                               mask=mask_[:, idx] if mask is not None else None,
                               mask_params=mask_params, axes=axes[idx],
                               outlines=outlines, image_mask=image_mask,
@@ -1352,6 +1420,9 @@ def plot_evoked_topomap(evoked, times="auto", ch_type=None, layout=None,
             cax.set_title(unit)
         cbar = fig.colorbar(images[-1], ax=cax, cax=cax, format=cbar_fmt)
         cbar.set_ticks([cbar.vmin, 0, cbar.vmax])
+        if cmap[1]:
+            for im in images:
+                im.axes.CB = DraggableColorbar(cbar, im)
 
     if proj == 'interactive':
         _check_delayed_ssp(evoked)
@@ -1366,8 +1437,8 @@ def plot_evoked_topomap(evoked, times="auto", ch_type=None, layout=None,
     return fig
 
 
-def _plot_topomap_multi_cbar(data, pos, ax, title=None, unit=None,
-                             vmin=None, vmax=None, cmap='RdBu_r',
+def _plot_topomap_multi_cbar(data, pos, ax, title=None, unit=None, vmin=None,
+                             vmax=None, cmap=None, outlines='head',
                              colorbar=False, cbar_fmt='%3.3f'):
     """Aux Function"""
     import matplotlib.pyplot as plt
@@ -1377,11 +1448,15 @@ def _plot_topomap_multi_cbar(data, pos, ax, title=None, unit=None,
     vmin = np.min(data) if vmin is None else vmin
     vmax = np.max(data) if vmax is None else vmax
 
+    if cmap == 'interactive':
+        cmap = (None, True)
+    elif not isinstance(cmap, tuple):
+        cmap = (cmap, True)
     if title is not None:
         ax.set_title(title, fontsize=10)
     im, _ = plot_topomap(data, pos, vmin=vmin, vmax=vmax, axes=ax,
-                         cmap=cmap, image_interp='bilinear', contours=False,
-                         show=False)
+                         cmap=cmap[0], image_interp='bilinear', contours=False,
+                         outlines=outlines, show=False)
 
     if colorbar is True:
         divider = make_axes_locatable(ax)
@@ -1391,6 +1466,8 @@ def _plot_topomap_multi_cbar(data, pos, ax, title=None, unit=None,
         if unit is not None:
             cbar.ax.set_title(unit, fontsize=8)
         cbar.ax.tick_params(labelsize=8)
+        if cmap[1]:
+            ax.CB = DraggableColorbar(cbar, im)
 
 
 @verbose
@@ -1452,9 +1529,16 @@ def plot_epochs_psd_topomap(epochs, bands=None, vmin=None, vmax=None,
         file is inferred from the data; if no appropriate layout file was
         found, the layout is automatically generated from the sensor
         locations.
-    cmap : matplotlib colormap
-        Colormap. For magnetometers and eeg defaults to 'RdBu_r', else
-        'Reds'.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode the colors are adjustable by clicking and dragging the
+        colorbar with left and right mouse button. Left mouse button moves the
+        scale up and down and right mouse button adjusts the range. Hitting
+        space bar resets the range. Up and down arrows can be used to change
+        the colormap. If None (default), 'Reds' is used for all positive data,
+        otherwise defaults to 'RdBu_r'. If 'interactive', translates to
+        (None, True).
     agg_fun : callable
         The function used to aggregate over frequencies.
         Defaults to np.sum. if normalize is True, else np.mean.
@@ -1515,8 +1599,8 @@ def plot_epochs_psd_topomap(epochs, bands=None, vmin=None, vmax=None,
 
 def plot_psds_topomap(
         psds, freqs, pos, agg_fun=None, vmin=None, vmax=None, bands=None,
-        cmap='RdBu_r', dB=True, normalize=False, cbar_fmt='%0.3f',
-        outlines='head', show=True):
+        cmap=None, dB=True, normalize=False, cbar_fmt='%0.3f', outlines='head',
+        show=True):
     """Plot spatial maps of PSDs
 
     Parameters
@@ -1545,9 +1629,16 @@ def plot_psds_topomap(
             bands = [(0, 4, 'Delta'), (4, 8, 'Theta'), (8, 12, 'Alpha'),
                      (12, 30, 'Beta'), (30, 45, 'Gamma')]
 
-    cmap : matplotlib colormap
-        Colormap. For magnetometers and eeg defaults to 'RdBu_r', else
-        'Reds'.
+    cmap : matplotlib colormap | (colormap, bool) | 'interactive' | None
+        Colormap to use. If tuple, the first value indicates the colormap to
+        use and the second value is a boolean defining interactivity. In
+        interactive mode the colors are adjustable by clicking and dragging the
+        colorbar with left and right mouse button. Left mouse button moves the
+        scale up and down and right mouse button adjusts the range. Hitting
+        space bar resets the range. Up and down arrows can be used to change
+        the colormap. If None (default), 'Reds' is used for all positive data,
+        otherwise defaults to 'RdBu_r'. If 'interactive', translates to
+        (None, True).
     dB : bool
         If True, transform data to decibels (with ``10 * np.log10(data)``)
         following the application of `agg_fun`. Only valid if normalize is
@@ -1607,8 +1698,8 @@ def plot_psds_topomap(
         else:
             unit = 'power'
 
-        _plot_topomap_multi_cbar(data, pos, ax, title=title,
-                                 vmin=vmin, vmax=vmax, cmap=cmap,
+        _plot_topomap_multi_cbar(data, pos, ax, title=title, vmin=vmin,
+                                 vmax=vmax, cmap=cmap, outlines=outlines,
                                  colorbar=True, unit=unit, cbar_fmt=cbar_fmt)
     tight_layout(fig=fig)
     fig.canvas.draw()
@@ -1680,7 +1771,6 @@ def _onselect(eclick, erelease, tfr, pos, ch_type, itmin, itmax, ifmin, ifmax,
         data = np.mean(data[indices, ifmin:ifmax, itmin:itmax], axis=0)
         chs = [tfr.ch_names[picks[x]] for x in indices]
     elif ch_type == 'grad':
-        picks = pick_types(tfr.info, meg=ch_type, ref_meg=False)
         from ..channels.layout import _pair_grad_sensors
         grads = _pair_grad_sensors(tfr.info, layout=layout,
                                    topomap_coords=False)
@@ -1700,8 +1790,14 @@ def _onselect(eclick, erelease, tfr, pos, ch_type, itmin, itmax, ifmin, ifmax,
     if not plt.fignum_exists(fig[0].number):
         fig[0] = figure_nobar()
     ax = fig[0].add_subplot(111)
-    itmax = min(itmax, len(tfr.times) - 1)
-    ifmax = min(ifmax, len(tfr.freqs) - 1)
+    itmax = len(tfr.times) - 1 if itmax is None else min(itmax,
+                                                         len(tfr.times) - 1)
+    ifmax = len(tfr.freqs) - 1 if ifmax is None else min(ifmax,
+                                                         len(tfr.freqs) - 1)
+    if itmin is None:
+        itmin = 0
+    if ifmin is None:
+        ifmin = 0
     extent = (tfr.times[itmin] * 1e3, tfr.times[itmax] * 1e3, tfr.freqs[ifmin],
               tfr.freqs[ifmax])
 
@@ -1733,8 +1829,9 @@ def _prepare_topomap(pos, ax):
 
 def _hide_frame(ax):
     """Helper to hide axis frame for topomaps."""
-    ax.set_xticks([])
-    ax.set_yticks([])
+    ax.get_yticks()
+    ax.xaxis.set_ticks([])
+    ax.yaxis.set_ticks([])
     ax.set_frame_on(False)
 
 
@@ -1980,8 +2077,7 @@ def _topomap_animation(evoked, ch_type='mag', times=None, frame_rate=None,
                                    frames=len(frames), interval=interval,
                                    blit=blit)
     fig.mne_animation = anim  # to make sure anim is not garbage collected
-    if show:
-        plt.show()
+    plt_show(show, block=False)
     if 'line' in params:
         # Finally remove the vertical line so it does not appear in saved fig.
         params['line'].remove()