+++ tutorials/plot_visualize_epochs.py
@@ -44,8 +44,12 @@ epochs.plot(block=True)
 # To plot individual channels as an image, where you see all the epochs at one
 # glance, you can use function :func:`mne.Epochs.plot_image`. It shows the
 # amplitude of the signal over all the epochs plus an average of the
-# activation.
-epochs.plot_image(97)
+# activation. We explicitly set interactive colorbar on (it is also on by
+# default for plotting functions with a colorbar except the topo plots). In
+# interactive mode you can scale and change the colormap with mouse scroll and
+# up/down arrow keys. You can also drag the colorbar with left/right mouse
+# button. Hitting space bar resets the scale.
+epochs.plot_image(97, cmap='interactive')
 
 # You also have functions for plotting channelwise information arranged into a
 # shape of the channel array. The image plotting uses automatic scaling by